backend/requirements.txt
pytest>=8.0.0
pytest-cov>=4.1.0
pydantic>=2.10.0
SQLAlchemy>=2.0.25
alembic>=1.13.1
psycopg[binary]>=3.2.12
openai>=1.12.0
qdrant-client>=1.15.0
httpx>=0.28.0
fastapi>=0.115.0
uvicorn>=0.32.0
python-dotenv>=1.2.0
python-multipart>=0.0.6
cryptography>=41.0.0
docker>=7.0.0
pytest-asyncio>=0.23.0
aiofiles>=23.2.1
